Fwd: Final Version of Express Enrollment
We need to create a custom bi-lingual kentico form for placing it on the
enrolment pages:

http://www.cryo-cell.com/enroll and http://www.cryo-cell.com/enroll-es
See attachment for the proposed layout.

Also,
For our current mobile http://www.cryo-cell.com/mobile/contact contact
form, We need to replace the code that sends the request to unbounce
with “the same” code we use now use on
http://www.cryo-cell.com/request-information that send information
directly toSalesForce.com instead. I suspect that this is a minor tweak
to the code behind the mobile contact form.

Regarding the  current http://www.cryo-cell.com/request-information form
implementation.
We need to add some logic there to pass a language used on the form to
Salesforce as a new custom field and also after the form post to
redirect to a localized thank you page.
As of now it redirects to /resources/request-forms/thank-you regardless
of the language on the form. We need to redirect it to
/resources/request-forms/thank-you-es    or
/resources/request-forms/thank-you-ru
if Spanish or Russian is selected on the /request-information form.